+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 ## [2.3.2] - not released yet
 ### Added
 - `FPDF.set_xmp_metadata`
+### Fixed
+- `FPDF.multi_cell`: line wrapping with justified content and unicode fonts, _cf._ [#118](https://github.com/PyFPDF/fpdf2/issues/118)
 
 ## [2.3.1] - 2021-02-28
 ### Added