+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 - made `<li>` indentation configurable through `fpdf.HTML2FPDF.LI_TAG_INDENT` class constant
 ### Fixed
 - `FPDF.multi_cell`: line wrapping with justified content and unicode fonts, _cf._ [#118](https://github.com/PyFPDF/fpdf2/issues/118)
+- `FPDF.multi_cell`: when `ln=3`, automatic page breaks now behave correctly at the bottom of pages
 
 ## [2.3.1] - 2021-02-28
 ### Added